pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spectre"
version = "0.1.0"
description = "Security Platform for Enrichment, Collection, Threat Research & Evaluation - An Agentic OSINT & Cyber Threat Intelligence Platform"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "SPECTRE Contributors" }
]
keywords = [
    "osint",
    "threat-intelligence",
    "cybersecurity",
    "reconnaissance",
    "mitre-attack",
    "stix",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Typing :: Typed",
]
dependencies = [
    # Core
    "httpx>=0.27.0",
    "pydantic[email]>=2.6.0",
    "pydantic-settings>=2.2.0",

    # CLI
    "typer>=0.12.0",
    "rich>=13.7.0",

    # DNS & WHOIS
    "dnspython>=2.6.0",
    "python-whois>=0.9.4",

    # Logging
    "structlog>=24.1.0",

    # Async
    "anyio>=4.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

# Future threat intel dependencies
threat-intel = [
    "stix2>=3.0.1",
    "taxii2-client>=2.3.0",
    "mitreattack-python>=3.0.0",
]

# Future graph analysis
graph = [
    "networkx>=3.2.0",
]

# All optional dependencies
all = [
    "spectre[dev,threat-intel,graph]",
]

[project.scripts]
spectre = "spectre.cli.main:app"

[project.entry-points."spectre.plugins"]
dns_recon = "spectre.plugins.osint.dns_recon:DNSReconPlugin"
whois_lookup = "spectre.plugins.osint.whois_lookup:WhoisLookupPlugin"

[project.urls]
Homepage = "https://github.com/spectre-osint/spectre"
Documentation = "https://github.com/spectre-osint/spectre#readme"
Repository = "https://github.com/spectre-osint/spectre"
Issues = "https://github.com/spectre-osint/spectre/issues"

[tool.hatch.build.targets.wheel]
packages = ["spectre"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise without from inside except
]

[tool.ruff.lint.isort]
known-first-party = ["spectre"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]
markers = [
    "integration: marks tests as integration tests (require network)",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_any_generics = true

[[tool.mypy.overrides]]
module = [
    "whois.*",
    "dns.*",
]
ignore_missing_imports = true

[tool.coverage.run]
source = ["spectre"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
